fix memory leak when building a docker image: (#114)
When building a docker image, we need to send the docker context to the docker server. 
The docker context is all the files, that are usually in the same directory as the dockerfile recursively plus any other directory(also recursively) specified by a user. 
ie when you run;
`docker build -t myImage .` 
the dot at the end says that the current directory is the docker context. All the files in the current directory and any sub-directories have to be tarred(zipped) and sent to the docker server. 

If a specified docker context had a lot of files or very huge files; meli would leak memory because it would have to allocate memory(bytes) for each file.

We fixed this by creating a pool of bytes that we re-use for this operation.
See issue: https://github.com/komuw/meli/issues/113

You can create profile files;
  `go build -o meli cli/cli.go && ./meli --memprofile mem.prof -up -f testdata/docker-compose.yml -d`  
and then use pprof(github.com/google/pprof or go tool pprof) 
  `pprof -sample_index=alloc_space meli mem.prof`
inside the pprof session;
  (pprof) top8 -cum
  (pprof) web
  (pprof) web walkFnClosure

We also fixed the build docker image benchmark so that you can now do:
  `go test -memprofile mem.prof -bench=BenchmarkBuildDockerImage -run=XXX .`
and pprof;
  `pprof -sample_index=alloc_space meli.test mem.prof`

read; 
1. https://austburn.me/blog/go-profile.html
2. https://github.com/golang/go/wiki/Performance
3. https://blog.golang.org/profiling-go-programs
4. https://golang.org/pkg/runtime/pprof